improved logging in graph api
