+++ internal/http/services/owncloud/ocgraph/conversions.go
@@ -2,6 +2,7 @@ package ocgraph
 
 import (
 	"context"
+	"fmt"
 	"path"
 	"path/filepath"
 	"time"
@@ -438,5 +439,45 @@ func (s *svc) toGrantee(ctx context.Context, recipientType string, id string) (*
 		}, nil
 	default:
 		return nil, errors.New(recipientType + " is not a valid granteetype")
+	}
+}
+
+func (s *svc) convertShareToSpace(rsi *gateway.ReceivedShareResourceInfo) *libregraph.Drive {
+	// the prefix of the remote_item.id and rootid
+	spacePath, _ := spaces.ResourceToSpacePath(rsi.ResourceInfo)
+	resourceRelativePath, _ := spaces.PathRelativeToSpaceRoot(rsi.ResourceInfo)
+
+	return &libregraph.Drive{
+		Id:         libregraph.PtrString(libregraphShareID(rsi.ReceivedShare.Share.Id)),
+		DriveType:  libregraph.PtrString("mountpoint"),
+		DriveAlias: libregraph.PtrString(rsi.ReceivedShare.Share.Id.OpaqueId), // this is not used, but must not be the same alias as the drive item
+		Name:       filepath.Base(rsi.ResourceInfo.Path),
+		WebUrl:     libregraph.PtrString(fullURL(s.c.WebBase, rsi.ResourceInfo.Path)),
+		Quota: &libregraph.Quota{
+			Total:     libregraph.PtrInt64(24154390300000),
+			Used:      libregraph.PtrInt64(3141592),
+			Remaining: libregraph.PtrInt64(24154387158408),
+		},
+		Root: &libregraph.DriveItem{
+			Id:        libregraph.PtrString(fmt.Sprintf("%s$%s!%s", ShareJailID, ShareJailID, rsi.ReceivedShare.Share.Id.OpaqueId)),
+			WebDavUrl: libregraph.PtrString(fullURL(s.c.WebDavBase, rsi.ResourceInfo.Path)),
+			RemoteItem: &libregraph.RemoteItem{
+				// Drive Alias does not contain the first '/'
+				DriveAlias: libregraph.PtrString(spacePath[1:]),
+				ETag:       libregraph.PtrString(rsi.ResourceInfo.Etag),
+				Folder:     &libregraph.Folder{},
+				// The Id must correspond to the id in the OCS response, for the time being
+				// It is in the form <something>!<something-else>
+				Id:                   libregraph.PtrString(spaces.EncodeResourceID(rsi.ResourceInfo.Id)),
+				LastModifiedDateTime: libregraph.PtrTime(time.Unix(int64(rsi.ResourceInfo.Mtime.Seconds), int64(rsi.ResourceInfo.Mtime.Nanos))),
+				Name:                 libregraph.PtrString(filepath.Base(rsi.ResourceInfo.Path)),
+				Path:                 libregraph.PtrString(resourceRelativePath),
+				// RootId must have the same token before ! as Id
+				// the second part for the time being is not used
+				RootId: libregraph.PtrString(fmt.Sprintf("%s$%s!unused_root_id", rsi.ResourceInfo.Id.StorageId, rsi.ResourceInfo.Id.SpaceId)),
+				Size:   libregraph.PtrInt64(int64(rsi.ResourceInfo.Size)),
+			},
+		},
+		Special: []libregraph.DriveItem{},
 	}
 }
